+++ example/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
 import 'package:timeframe_selector/timeframe_selector.dart';
 
 void main() async {
@@ -28,29 +29,70 @@ class TimeFrameExample extends StatefulWidget {
 class _TimeFrameExampleState extends State<TimeFrameExample> {
   DateTimeRange? selectedTimeframe;
 
+  String? getTimeframeString() {
+    if (selectedTimeframe != null) {
+      final startString = DateFormat.Hm().format(
+        selectedTimeframe!.start,
+      );
+      final endString = DateFormat.Hm().format(
+        selectedTimeframe!.end,
+      );
+      return "$startString - $endString";
+    }
+    return null;
+  }
+
   @override
   Widget build(BuildContext context) {
     return SingleChildScrollView(
-      child: TimeframeSelector(
-        baseTimeframe: DateTimeRange(
-          start: DateTime(0, 1, 1, 10),
-          end: DateTime(0, 1, 1, 24),
-        ),
-        noSelectionText: "No selection",
-        timeSegmentLength: const Duration(minutes: 30),
-        minTimeframeSegments: 4,
-        selectedTimeframe: selectedTimeframe,
-        blockedTimeframes: [
-          DateTimeRange(
-            start: DateTime(0, 1, 1, 15),
-            end: DateTime(0, 1, 1, 17),
+      child: Column(
+        children: [
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Text(getTimeframeString() ?? "No selection"),
+          ),
+          TimeframeSelector(
+            baseTimeframe: DateTimeRange(
+              start: DateTime(0, 1, 1, 10),
+              end: DateTime(0, 1, 1, 24),
+            ),
+            timeSegmentLength: const Duration(minutes: 30),
+            minTimeframeSegments: 4,
+            selectedTimeframe: selectedTimeframe,
+            overlayDatas: [
+              OverlayData(
+                timeframes: [
+                  DateTimeRange(
+                    start: DateTime(0, 1, 1, 15),
+                    end: DateTime(0, 1, 1, 17),
+                  ),
+                ],
+              ),
+              OverlayData(
+                timeframes: [
+                  DateTimeRange(
+                    start: DateTime(0, 1, 1, 20),
+                    end: DateTime(0, 1, 1, 22),
+                  ),
+                ],
+                color: Colors.green.withOpacity(0.2),
+                borderColor: Colors.green.withOpacity(0.4),
+                blockSelection: false,
+                child: const Text("Hello world!"),
+                onTap: () {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(content: Text("Hello world!")),
+                  );
+                },
+              ),
+            ],
+            onTimeframeChange: (newTimeframe) => setState(
+              () {
+                selectedTimeframe = newTimeframe;
+              },
+            ),
           ),
         ],
-        onTimeframeChange: (newTimeframe) => setState(
-          () {
-            selectedTimeframe = newTimeframe;
-          },
-        ),
       ),
     );
   }