+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.2.0
+
+**Breaking**
+* Removed "noSelectionText" parameter
+
+**Changes**
+* Removed selected timeframe text as it has no direct correlation to the timeframe selector
+
+## 0.1.0
+
+**Breaking**
+* Replaced blockedTimeframes from TimeframeSelector with overlayDatas
+
+**Changes**
+* Added OverlayData to allow customisation of overlays
+
+
 ## 0.0.1
 
 * First release of timeframe_selector